+++ src/meegsim/coupling_graph.py
@@ -81,18 +81,14 @@ def _set_coupling(sources, coupling_graph, times, random_state):
     Parameters
     ----------
     sources : dict
-        Simulated sources.
+        Simulated sources. Their waveforms are modified in-place by the
+        coupling function(s).
     coupling_graph : nx.Graph
         The coupling graph that describes the desired connectivity pattern.
     times : array-like
         The time points for all samples in the waveform.
     random_state : int or None
         The random state that could be fixed to ensure reproducibility.
-
-    Returns
-    -------
-    sources : dict
-        Simulated sources with waveforms adjusted according to the desired coupling.
     """
     walkaround = generate_walkaround(coupling_graph, random_state=random_state)
 
@@ -114,5 +110,3 @@ def _set_coupling(sources, coupling_graph, times, random_state):
             **tmp_coupling_params,
             random_state=random_state,
         )
-
-    return sources