+++ tests/test_snr.py
@@ -177,6 +177,7 @@ def test_adjust_snr_local_point(adjust_snr_mock):
             waveform=np.ones((1, 100)),
             snr=np.array([5.0]),
             snr_params=dict(fmin=8, fmax=12),
+            std=1,
             names=["s1"],
         ),
         PointSourceGroup(
@@ -185,6 +186,7 @@ def test_adjust_snr_local_point(adjust_snr_mock):
             waveform=np.ones((1, 100)),
             snr=None,
             snr_params=dict(),
+            std=1,
             names=["s2"],
         ),
     ]
@@ -195,7 +197,7 @@ def test_adjust_snr_local_point(adjust_snr_mock):
     noise_sources = {"n1": prepare_point_source(name="n1")}
     tstep = 0.01
 
-    sources = _adjust_snr_local(src, fwd, tstep, sources, source_groups, noise_sources)
+    _adjust_snr_local(src, fwd, tstep, sources, source_groups, noise_sources)
 
     # Check the SNR adjustment was performed only once
     adjust_snr_mock.assert_called_once()
@@ -218,6 +220,7 @@ def test_adjust_snr_local_patch(adjust_snr_mock):
             waveform=np.ones((1, 100)),
             snr=np.array([5.0]),
             snr_params=dict(fmin=8, fmax=12),
+            std=1,
             extents=None,
             names=["s1"],
         ),
@@ -227,6 +230,7 @@ def test_adjust_snr_local_patch(adjust_snr_mock):
             waveform=np.ones((1, 100)),
             snr=None,
             snr_params=dict(),
+            std=1,
             extents=None,
             names=["s2"],
         ),
@@ -238,7 +242,7 @@ def test_adjust_snr_local_patch(adjust_snr_mock):
     noise_sources = {"n1": prepare_point_source(name="n1")}
     tstep = 0.01
 
-    sources = _adjust_snr_local(src, fwd, tstep, sources, source_groups, noise_sources)
+    _adjust_snr_local(src, fwd, tstep, sources, source_groups, noise_sources)
 
     # Check the SNR adjustment was performed only once
     adjust_snr_mock.assert_called_once()
@@ -276,7 +280,7 @@ def test_adjust_snr_global_point(adjust_snr_mock):
     noise_sources = {"n1": prepare_point_source(name="n1")}
     tstep = 0.01
 
-    sources = _adjust_snr_global(
+    _adjust_snr_global(
         src,
         fwd,
         snr_global=5,
@@ -308,7 +312,7 @@ def test_adjust_snr_global_patch(adjust_snr_mock):
     noise_sources = {"n1": prepare_point_source(name="n1")}
     tstep = 0.01
 
-    sources = _adjust_snr_global(
+    _adjust_snr_global(
         src,
         fwd,
         snr_global=5,